requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.4
django-filter==1.1.0
django-rest-swagger==2.1.2
djangorestframework==3.8.2
idna==2.6
itypes==1.1.0
Jinja2==2.10
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
openapi-codec==1.3.2
pip==8.1.1
pkg-resources==0.0.0
Pygments==2.2.0
pytz==2018.4
requests==2.18.4
setuptools==20.7.0
simplejson==3.13.2
uritemplate==3.0.0
urllib3==1.22
